requirements.txt
astroid==2.9.2
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.13.0
django-environ==0.8.1
django-heroku==0.3.1
django-registration-redux==2.9
django-smtp-ssl==1.0
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.0
pip-check-reqs==2.3.2
psycopg2==2.9.3
pycparser==2.21
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
PyMySQL==1.0.2
pyparsing==3.0.6
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typed-ast==1.5.1
whitenoise==5.3.0
wincertstore==0.2
wrapt==1.13.3
